Rename question page file, adjust main page CSS
